--- src/main/java/com/ahnhaetdaeyo/finalbe/security/userDetail/UserDetail.java
@@ -1,4 +1,0 @@
-package com.ahnhaetdaeyo.finalbe.security.userDetail;
-
-public class UserDetail {
-}
