--- src/main/java/com/ahnhaetdaeyo/finalbe/security/userDetail/UserDetailService.java
@@ -1,4 +1,0 @@
-package com.ahnhaetdaeyo.finalbe.security.userDetail;
-
-public class UserDetailService {
-}
