+++ src/main/java/com/havit/finalbe/FinalBeApplication.java
@@ -1,4 +1,4 @@
-package com.ahnhaetdaeyo.finalbe;
+package com.havit.finalbe;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
