+++ src/main/java/com/havit/finalbe/configuration/SwaggerConfig.java
@@ -1,4 +1,4 @@
-package com.ahnhaetdaeyo.finalbe.configuration;
+package com.havit.finalbe.configuration;
 
 import io.swagger.v3.oas.annotations.OpenAPIDefinition;
 import io.swagger.v3.oas.annotations.info.Contact;
