+++ src/main/java/com/havit/finalbe/security/userDetail/UserDetail.java
@@ -1,0 +1,4 @@
+package com.havit.finalbe.security.userDetail;
+
+public class UserDetail {
+}
