+++ src/test/java/com/havit/finalbe/FinalBeApplicationTests.java
@@ -1,4 +1,4 @@
-package com.ahnhaetdaeyo.finalbe;
+package com.havit.finalbe;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
